Players can capture enemy buildings Use new keybind "capture", default C
Go to interact with the enemy building
and press and hold the capture key;
a capture bar should be rendered and
when it reaches 100% full the building
should be captured.

If two players try to capture the
building at the same time, neither
capture succeeds and any existing
capture progress is lost.

If a player capturing an enemy building
takes damage, then their capture progress
is lost.