Refactored and altered to improve performance
Improved DebugSynchedRandom to hold on to garbage
strings for less time, using a flywheel cache
of intern()ed strings.

Increased frequency of sync packets to 1 in 20
turns from 1 in 50, to reduce the memory overhead
(although the above change should probably mean
this is less important now).

Hard-coded maxMonsters to be 400, as higher values
cause very poor FPS (probably machine dependent).

Added debug prints to game loop to show number
of entities on the level.