LocalGameInput: ignore auto-repeat keypresses On Windows, at least
On Windows key events are repeatedly fired when
keys are held down, due to auto-repeat. The JVM
generates key-press and key-typed events for
each auto-repeat, but only generated a key-released
when the key is actually released.

So by ignoring key presses that occur while a key
is already marked as "down" we can ignore the
auto-repeats. Since we only want this behaviour
in-game (keep auto-repeat for menus for normal
input), it only ignores key presses for logical
input updates.

This may only work on Windows, because the I read
that the JVM generates the auto-repeat events
differently on Linux, but I haven't been able to
test this yet. (Rumour is that Linux auto-repeats
generate key-releases too.)