Refactor menu background render & in-game opacity
Default background now rendered in the GameView.
Only a few menus differ and blit over with the
Art.emptyBackground.

Change opacity of background for all menus rendered
while the game is running (back to how it used to
be -- I think the change was an unintentional
consequence of changes to the way alpha values are
passed in Bitmap#alphaBlit() and Bitmap#alphaFill()).