Fix multiplayer sync issue and clean up a bit
Sync issue was due to one Player being
added to Level at a different point on
the server and client so that they were
in a different index in the entity list
leading to logic differences.

Removed some debug prints

Added a quiet option for sync_check --
only prints output when the game gets
out of sync.

NB: This implementation of sync check
    differs from the sync check branch
    in not limiting the number of entities
    which slows the game down a lot and
    produces large amounts in the syncdump
    files if game goes out of sync.

DebugSynchedRandom: changed count field
to be private; added getCount() method